fix: split-interface-tools
