+++ packages/SwingSet/src/liveslots/virtualObjectManager.js
@@ -2,12 +2,11 @@
 
 import { assert, Fail } from '@agoric/assert';
 import { objectMap } from '@agoric/internal';
+import { assertPattern, mustMatch } from '@agoric/store';
 import {
-  assertPattern,
   defendPrototype,
   defendPrototypeKit,
-  mustMatch,
-} from '@agoric/store';
+} from '@agoric/store/exo-tools.js';
 import { Far, hasOwnPropertyOf, passStyleOf } from '@endo/marshal';
 import { parseVatSlot } from '../lib/parseVatSlots.js';
 import { enumerateKeysWithPrefix } from './vatstore-iterators.js';
