+++ packages/store/exo-tools.js
@@ -1,0 +1,4 @@
+export {
+  defendPrototype,
+  defendPrototypeKit,
+} from './src/patterns/interface-tools.js';
