+++ packages/store/package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "@agoric/assert": "^0.5.1",
     "@agoric/internal": "^0.2.1",
+    "@endo/far": "^0.2.14",
     "@endo/eventual-send": "^0.16.8",
     "@endo/marshal": "^0.8.1",
     "@endo/promise-kit": "^0.2.52",
