+++ packages/store/src/patterns/interface-tools.js
@@ -1,12 +1,11 @@
-import { Far } from '@endo/marshal';
-import { E } from '@endo/eventual-send';
+import { E, Far } from '@endo/far';
 import { listDifference, objectMap } from '@agoric/internal';
 
 import { MinMethodGuard, mustMatch, M } from './patternMatchers.js';
 
 const { quote: q, Fail } = assert;
 const { apply, ownKeys } = Reflect;
-const { defineProperties, seal, freeze } = Object;
+const { defineProperties } = Object;
 
 const defendSyncArgs = (args, methodGuard, label) => {
   const { argGuards, optionalArgGuards, restArgGuard } = methodGuard;
@@ -273,163 +272,3 @@ export const defendPrototypeKit = (
     ),
   );
 };
-
-const emptyRecord = harden({});
-
-/**
- * When calling `defineDurableKind` and
- * its siblings, used as the `init` function argument to indicate that the
- * state record of the (virtual/durable) instances of the kind/exoClass
- * should be empty, and that the returned maker function should have zero
- * parameters.
- *
- * @returns {{}}
- */
-export const initEmpty = () => emptyRecord;
-
-/**
- * @template [S = any]
- * @template [T = any]
- * @typedef {object} Context
- * @property {S} state
- * @property {T} self
- */
-
-/**
- * @template [S = any]
- * @template [F = any]
- * @typedef {object} KitContext
- * @property {S} state
- * @property {F} facets
- */
-
-/**
- * @typedef {{[name: string]: Pattern}} StateShape
- * It looks like a copyRecord pattern, but the interpretation is different.
- * Each property is distinct, is checked and changed separately.
- */
-
-/**
- * @template C
- * @typedef {object} FarClassOptions
- * @property {(context: C) => void} [finish]
- * @property {StateShape} [stateShape]
- */
-
-/**
- * @template A args to init
- * @template S state from init
- * @template {Record<string | symbol, CallableFunction>} T methods
- * @param {string} tag
- * @param {any} interfaceGuard
- * @param {(...args: A[]) => S} init
- * @param {T & ThisType<{ self: T, state: S }>} methods
- * @param {FarClassOptions<Context<S,T>>} [options]
- * @returns {(...args: A[]) => (T & import('@endo/eventual-send').RemotableBrand<{}, T>)}
- */
-export const defineExoClass = (
-  tag,
-  interfaceGuard,
-  init,
-  methods,
-  { finish = undefined } = {},
-) => {
-  /** @type {WeakMap<T,Context<S, T>>} */
-  const contextMap = new WeakMap();
-  const prototype = defendPrototype(
-    tag,
-    contextMap,
-    methods,
-    true,
-    interfaceGuard,
-  );
-  const makeInstance = (...args) => {
-    // Be careful not to freeze the state record
-    const state = seal(init(...args));
-    /** @type {T} */
-    // @ts-expect-error could be instantiated with different subtype
-    const self = harden({ __proto__: prototype });
-    // Be careful not to freeze the state record
-    /** @type {Context<S,T>} */
-    const context = freeze({ state, self });
-    contextMap.set(self, context);
-    if (finish) {
-      finish(context);
-    }
-    return self;
-  };
-  // @ts-expect-error could be instantiated with different subtype
-  return harden(makeInstance);
-};
-harden(defineExoClass);
-
-/**
- * @template A args to init
- * @template S state from init
- * @template {Record<string, Record<string | symbol, CallableFunction>>} F methods
- * @param {string} tag
- * @param {any} interfaceGuardKit
- * @param {(...args: A[]) => S} init
- * @param {F & ThisType<{ facets: F, state: S }> } methodsKit
- * @param {FarClassOptions<KitContext<S,F>>} [options]
- * @returns {(...args: A[]) => F}
- */
-export const defineExoClassKit = (
-  tag,
-  interfaceGuardKit,
-  init,
-  methodsKit,
-  { finish = undefined } = {},
-) => {
-  const contextMapKit = objectMap(methodsKit, () => new WeakMap());
-  const prototypeKit = defendPrototypeKit(
-    tag,
-    contextMapKit,
-    methodsKit,
-    true,
-    interfaceGuardKit,
-  );
-  const makeInstanceKit = (...args) => {
-    // Be careful not to freeze the state record
-    const state = seal(init(...args));
-    // Don't freeze context until we add facets
-    const context = { state };
-    const facets = objectMap(prototypeKit, (prototype, facetName) => {
-      const self = harden({ __proto__: prototype });
-      contextMapKit[facetName].set(self, context);
-      return self;
-    });
-    context.facets = facets;
-    // Be careful not to freeze the state record
-    freeze(context);
-    if (finish) {
-      // @ts-expect-error `facets` was added
-      finish(context);
-    }
-    return facets;
-  };
-  // eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error -- different per package https://github.com/Agoric/agoric-sdk/issues/4620
-  // @ts-ignore xxx
-  return harden(makeInstanceKit);
-};
-harden(defineExoClassKit);
-
-/**
- * @template {Record<string, Method>} T
- * @param {string} tag
- * @param {InterfaceGuard | undefined} interfaceGuard CAVEAT: static typing does not yet support `callWhen` transformation
- * @param {T} methods
- * @param {FarClassOptions<Context<{},T>>} [options]
- * @returns {T & import('@endo/eventual-send').RemotableBrand<{}, T>}
- */
-export const makeExo = (tag, interfaceGuard, methods, options = undefined) => {
-  const makeInstance = defineExoClass(
-    tag,
-    interfaceGuard,
-    initEmpty,
-    methods,
-    options,
-  );
-  return makeInstance();
-};
-harden(makeExo);
